[LR11x0] Added explicit typecast

+++ src/modules/LR11x0/LR11x0.cpp
@@ -358,7 +358,7 @@ int16_t LR11x0::standby(uint8_t mode, bool wakeup) {
   if(wakeup) {
     // send a NOP command - this pulls the NSS low to exit the sleep mode,
     // while preventing interference with possible other SPI transactions
-    (void)this->mod->SPIwriteStream(RADIOLIB_LR11X0_CMD_NOP, NULL, 0, false, false);
+    (void)this->mod->SPIwriteStream((uint16_t)RADIOLIB_LR11X0_CMD_NOP, NULL, 0, false, false);
   }
 
   uint8_t buff[] = { mode };
